refactor: move CredencialesDTO to dto package
