+++ src/main/java/com/example/restapi/dto/CredencialesDTO.java
@@ -1,0 +1,41 @@
+package com.example.restapi.dto;
+
+public class CredencialesDTO {
+
+    private String usuario;
+    private String contrasena;
+
+    
+    public CredencialesDTO() {
+    }
+
+    public CredencialesDTO(String usuario, String contrasena) {
+        this.usuario = usuario;
+        this.contrasena = contrasena;
+    }
+
+  
+    public String getUsuario() {
+        return usuario;
+    }
+
+    public void setUsuario(String usuario) {
+        this.usuario = usuario;
+    }
+
+    public String getContrasena() {
+        return contrasena;
+    }
+
+    public void setContrasena(String contrasena) {
+        this.contrasena = contrasena;
+    }
+
+    @Override
+    public String toString() {
+        return "CredencialesDTO{" +
+                "usuario='" + usuario + '\'' +
+                ", contrasena='" + contrasena + '\'' +
+                '}';
+    }
+}
